requirements.txt
streamlit
google-genai
google-cloud-aiplatform
pandas
python-dotenv
tornado==6.3.3
google-generativeai>=0.4.0
vertexai
google-cloud-firestore